이름들을 저장하는 기능을 Names 클래스로 분리했다.
이름들을 List로 저장하고 출력 시 이름들을 하나의 row로 만들어주는 기능을 분리해서 클래스로 만들었다.
